# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Bakery API
  description: >
    Order management for the bakery: shop заявки, baker matrix, production
    sheets (отработка), dough monitoring, and admin dictionaries. All error
    responses use the `{"error": "<safe message>"}` envelope.
  version: 1.0.0
servers:
  - url: /
    description: Direct backend access
...
security:
  - bearer: []
  - tma: []
paths:
...
  /orders/{id}/favorite:
    patch:
      tags: [admin]
      summary: Toggle the favorite flag (admin)
      operationId: setOrderFavorite
      parameters: [{ $ref: "#/components/parameters/OrderNumber" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [favorite]
              properties:
                favorite: { type: boolean }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /admin/dishes:
    get:
      tags: [admin]
      summary: Dish catalog with codes (admin)
      operationId: listDishes
      responses:
        "200":
          description: Catalog rows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Dish" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [admin]
      summary: Add a dish to the catalog
      operationId: createDish
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DishWrite" }
      responses:
        "200":
          description: Created dish
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dish" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/dishes/techcards:
    get:
      tags: [admin]
      summary: Dish catalog tech cards, grouped by order category
      operationId: listDishTechCards
      responses:
        "200":
          description: Categories with each dish's tech card ingredients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TechCardCategory" }
        "403": { $ref: "#/components/responses/Error" }
  /admin/dishes/available:
    get:
      tags: [admin]
      summary: Search iiko DISH products to add
      operationId: searchAvailableDishes
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Name or code substring }
      responses:
        "200":
          description: Matching iiko products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AvailableDish" }
        "403": { $ref: "#/components/responses/Error" }
...
  /admin/dishes/reorder:
    put:
      tags: [admin]
      summary: Reorder the whole catalog
      operationId: reorderDishes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codes]
              properties:
                codes:
                  type: array
                  items: { type: string }
      responses:
        "204": { description: Reordered }
        "400": { $ref: "#/components/responses/Error" }
  /admin/dishes/{code}:
    put:
      tags: [admin]
      summary: Update a catalog dish
      operationId: updateDish
      parameters: [{ $ref: "#/components/parameters/DishCode" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DishWrite" }
      responses:
        "200":
          description: Updated dish
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dish" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [admin]
      summary: Remove a dish from the catalog
      operationId: deleteDish
      parameters: [{ $ref: "#/components/parameters/DishCode" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
  /admin/categories:
    post:
      tags: [admin]
      summary: Create an order category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryWrite" }
      responses:
        "200":
          description: Created category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/categories/{id}:
    put:
      tags: [admin]
      summary: Update a category (code is immutable)
      operationId: updateCategory
      parameters: [{ $ref: "#/components/parameters/CategoryID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryWrite" }
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [admin]
      summary: Delete a category (fails while dishes reference it)
      operationId: deleteCategory
      parameters: [{ $ref: "#/components/parameters/CategoryID" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: HMAC web session token from POST /login (7-day TTL).
    tma:
      type: apiKey
      in: header
      name: Authorization
      description: '`tma <initData>` — Telegram Mini App initData, validated per request (24h max age).'
  parameters:
    OrderNumber:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Order number, e.g. Г.Х.08.07.26.001
...
    DishCode:
      name: code
      in: path
      required: true
      schema: { type: string }
    CategoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    Error:
      description: Error with a safe user-facing message
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
    Department:
      type: object
      required: [id, code, name, type]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: gagarina }
        name: { type: string, example: Магазин Гагарина }
        type: { type: string, enum: [shop, workshop] }
...
    Category:
      type: object
      required: [id, code, letter, name, color, sort_order, monitor_codes]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: bread }
        letter: { type: string, example: Х, description: Becomes part of the order number }
        name: { type: string, example: Хлеб }
        color:
          type: string
          enum: [amber, sky, violet, emerald, rose, stone]
          description: Must stay in sync with domain.CategoryColors and frontend lib/categories.js
        sort_order: { type: integer, format: int64 }
        monitor_codes:
          type: array
          items: { type: string }
          description: Dough codes used by /monitor for this category
    CategoryWrite:
      type: object
      required: [code, letter, name, color]
      properties:
        code: { type: string }
        letter: { type: string }
        name: { type: string }
        color: { type: string, enum: [amber, sky, violet, emerald, rose, stone] }
        sort_order: { type: integer, format: int64 }
        monitor_codes:
          type: array
          items: { type: string }
    OrderItem:
      type: object
      required: [code, product_name, quantity, reserved_quantity, production_quantity]
      properties:
        code: { type: string }
        product_name: { type: string }
        quantity: { type: number }
        reserved_quantity: { type: number }
        production_quantity: { type: number, description: quantity + reserved_quantity }
        produced_quantity:
          type: number
          nullable: true
          description: Production fact; absent = baked as ordered
        produced_reason: { type: string }
    Comments:
      type: object
      required: [general, items]
      properties:
        general: { type: string }
        items:
          type: array
          items:
            type: object
            required: [product_name, comment]
            properties:
              product_name: { type: string }
              comment: { type: string }
    HistoryItem:
      type: object
      required: [change_type, product_code, product_name]
      properties:
        change_type: { type: string, enum: [added, updated, removed] }
        product_code: { type: string }
        product_name: { type: string }
        old_quantity: { type: number, nullable: true }
        new_quantity: { type: number, nullable: true }
        old_reserved_quantity: { type: number, nullable: true }
        new_reserved_quantity: { type: number, nullable: true }
    HistoryEntry:
      type: object
      required: [id, changed_by_username, changed_at, items]
      properties:
        id: { type: integer, format: int64 }
        changed_by_username: { type: string }
        changed_at: { type: string, description: RFC3339 }
        items:
          type: array
          items: { $ref: "#/components/schemas/HistoryItem" }
    Order:
      type: object
      required: [id, number, location, created_by_username, items, created_at, fulfillment_date, monitor_command, comments, favorite, cancelled]
      properties:
        id: { type: string }
        number: { type: string, example: Г.Х.08.07.26.001 }
        location: { type: string }
        created_by_username: { type: string }
        from_department: { $ref: "#/components/schemas/Department" }
        to_department: { $ref: "#/components/schemas/Department" }
        category: { $ref: "#/components/schemas/Category" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        created_at: { type: string, description: RFC3339 }
        fulfillment_date: { type: string, format: date }
        monitor_command: { type: string, example: /monitor Г.Х.08.07.26.001 }
        comments: { $ref: "#/components/schemas/Comments" }
        favorite: { type: boolean }
        cancelled: { type: boolean }
        cancelled_by_username: { type: string }
        production_sheet_id:
          type: integer
          format: int64
          nullable: true
          description: The sheet covering this order (an order has at most one)
        history:
          type: array
          items: { $ref: "#/components/schemas/HistoryEntry" }
...
    DishCatalogItem:
      type: object
      required: [code, name, theme, sort_order]
      properties:
        code: { type: string }
        name: { type: string }
        theme: { type: string, description: "Group name, e.g. КОКРОКИ (normalized to a dish_groups row keyed by category+name)" }
        category_id: { type: integer, format: int64, nullable: true }
        sort_order: { type: integer, format: int64 }
    Dish:
      allOf:
        - $ref: "#/components/schemas/DishCatalogItem"
    DishWrite:
      type: object
      required: [code, name, theme]
      properties:
        code: { type: string, description: iiko product code }
        name: { type: string }
        theme: { type: string, description: "Group name; find-or-created in dish_groups under category_id" }
        category_id: { type: integer, format: int64, nullable: true }
        sort_order: { type: integer, format: int64 }
    AvailableDish:
      type: object
      required: [code, name, unit]
      properties:
        code: { type: string }
        name: { type: string }
        unit: { type: string }
    TechCardIngredient:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        unit: { type: string }
        amount: { type: number }
    TechCardDish:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        unit: { type: string }
        ingredients:
          type: array
          items: { $ref: "#/components/schemas/TechCardIngredient" }
        error: { type: string, description: Set instead of ingredients when the tech card could not be loaded }
    TechCardCategory:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        dishes:
          type: array
          items: { $ref: "#/components/schemas/TechCardDish" }
...